Allow the search callback to return an object with a `toArray` method
This adds out-of-the-box support for things like Ember Data promises.
Since a `toArray` method is very common on collection-like implementation
this will potentially also support other use cases.

If your search promise does not return an array or an object with a
`toArray` method, than you can still do the conversion manually like so:

```
customSearch(term) {
  return somePromise.then(results => {
    let array = convertToArray(results); // convert results to array
    return array;
  });
}
```