[RFC] Search should work when options is a promise
I've run into problems with searching options when options is a promise, which I did not have when using EPS itself. 

I believe this is addressed in EPS by making `options` a computed property and having it's computed `set` update if options is `thennable`.

This is an example of a fix to demonstrate the issue, but there are plenty of ways to patch this. I'm not sure which way you'd prefer, or if you'd prefer a more holistic fix.

Thanks!